requirements.txt
# --- Telegram ---
telethon==1.36.0
python-telegram-bot==21.7

# --- Web Panel ---
fastapi==0.115.2
uvicorn==0.30.6
jinja2==3.1.4

# --- Utilities ---
python-dotenv==1.0.1
apscheduler==3.10.4
pandas==2.2.3
matplotlib==3.9.2
python-multipart
httpx
pytz
